src/requirements.txt
autopep8==1.5.7
beautifulsoup4==4.10.0
black==21.9b0
bs4==0.0.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
click==8.0.3
colorama==0.4.4
commonmark==0.9.1
coreapi==2.3.3
coreschema==0.0.4
cryptography==36.0.1
Django==2.2.2
django-extensions==3.1.3
django-rest-swagger==2.2.0
django-simple-history==3.0.0
djangorestframework==3.12.4
et-xmlfile==1.1.0
flake8==3.9.2
future==0.18.2
idna==3.2
itypes==1.2.0
Jinja2==3.0.2
MarkupSafe==2.0.1
mccabe==0.6.1
mypy-extensions==0.4.3
openapi-codec==1.3.2
openpyxl==3.0.9
pathspec==0.9.0
pefile==2021.5.24
Pillow==8.3.2
platformdirs==2.4.0
psycopg2==2.9.1
pycodestyle==2.7.0
pycparser==2.21
pyflakes==2.3.1
Pygments==2.10.0
PyMuPDF==1.18.14
pyOpenSSL==21.0.0
pytz==2021.1
qrcode==7.3.1
regex==2021.10.23
reportlab==3.5.67
requests==2.26.0
rich==10.12.0
simplejson==3.17.5
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.2
toml==0.10.2
tomli==1.2.2
typing-extensions==3.10.0.2
Unidecode==1.3.2
uritemplate==4.1.1
urllib3==1.26.7
Werkzeug==2.0.2
xlrd==2.0.1
xlwt==1.3.0
